+++ src/trio_util/_async_value.py
@@ -3,12 +3,37 @@ from typing import TypeVar, Generic, AsyncIterator, Tuple, ContextManager, Calla
 
 import trio
 
-try:
-    from trio.lowlevel import ParkingLot as WaitQueue
-except ImportError:  # pragma: no cover
-    from trio.hazmat import ParkingLot as WaitQueue  # type: ignore[no-redef]
-
 from ._ref_counted_default_dict import _RefCountedDefaultDict
+
+
+class _WaitQueue:
+    """Fast replacement for trio.ParkingLot.
+
+    The typical unpark_all() is 2x faster.
+    """
+
+    __slots__ = ['tasks']
+
+    def __init__(self):
+        self.tasks = set()
+
+    def __len__(self):
+        return len(self.tasks)
+
+    async def park(self):
+        task = trio.lowlevel.current_task()
+        self.tasks.add(task)
+
+        def abort_fn(_):
+            self.tasks.remove(task)
+            return trio.lowlevel.Abort.SUCCEEDED
+
+        await trio.lowlevel.wait_task_rescheduled(abort_fn)
+
+    def unpark_all(self):
+        for task in self.tasks:
+            trio.lowlevel.reschedule(task)
+        self.tasks.clear()
 
 
 def _ANY_TRANSITION(value, old_value):
@@ -30,7 +55,7 @@ class _Result:
     __slots__ = ['event', 'value']
 
     def __init__(self):
-        self.event = WaitQueue()
+        self.event = _WaitQueue()
         self.value = None