+++ CHANGELOG.md
@@ -5,6 +5,9 @@
 - `AsyncValue` and `compose_values()` now have type hints and generic typing
 ### Changed
 - `_transform_` parameter of `compose_values()` is now keyword only
+### Fixed
+- slight performance improvement to `AsyncValue.value` setter when there is a
+  predicate match.
 
 ## trio-util 0.5.0 (2021-04-06)
 ⚠ includes breaking changes!